src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/get-distributed-nodes-orphan-expected.txt

CONSOLE ERROR: line 26: Uncaught TypeError: host.createShadowRoot is not a function
getDistributedNodes() should work out of Document

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
